# openapi_spec.yaml
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    HealthResponse:
      properties:
        response:
          title: Response
          type: string
        status:
          title: Status
          type: string
      required:
      - status
      - response
      title: HealthResponse
      type: object
    PingResponse:
      properties:
        response:
          title: Response
          type: string
      required:
      - response
      title: PingResponse
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    WordleSolveRequest:
      properties:
        places:
          items:
            type: string
          title: Places
          type: array
        size:
          title: Size
          type: integer
        words:
          items:
            type: string
          title: Words
          type: array
      required:
      - words
      - places
      - size
      title: WordleSolveRequest
      type: object
    WordleSolveResponse:
      properties:
        alternatives:
          items:
            type: string
          title: Alternatives
          type: array
        remaining:
          title: Remaining
          type: integer
        suggestions:
          items:
            type: string
          title: Suggestions
          type: array
        used_letters:
          items:
            type: string
          title: Used Letters
          type: array
      required:
      - suggestions
      - alternatives
      - remaining
      - used_letters
      title: WordleSolveResponse
      type: object
info:
  description: This is an API for Wordle
  title: Wordle API
  version: 1.0.0
openapi: 3.1.0
paths:
  /wordle-solver/api/health:
    get:
      operationId: health
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
          description: Successful Response
      summary: Health
  /wordle-solver/api/ping:
    get:
      operationId: ping
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PingResponse'
          description: Successful Response
      summary: Ping
  /wordle-solver/api/wordle:
    post:
      operationId: solve_wordle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WordleSolveRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WordleSolveResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Solve Wordle